MGMT-9176 Complete scenarios for networkType selection
