+++ src/common/components/clusterConfiguration/AdvancedNetworkFields.tsx
@@ -1,14 +1,19 @@
-import React, { useEffect, useMemo } from 'react';
+import React from 'react';
 import { Grid, TextInputTypes } from '@patternfly/react-core';
 import { useFormikContext } from 'formik';
 import { InputField } from '../../components/ui';
 import { NetworkConfigurationValues } from '../../types/clusters';
-import { Address6 } from 'ip-address';
-import { NETWORK_TYPE_OVN, PREFIX_MAX_RESTRICTION } from '../../config/constants';
+import { PREFIX_MAX_RESTRICTION } from '../../config/constants';
 import { NetworkTypeControlGroup } from '../clusterWizard/networkingSteps/NetworkTypeControlGroup';
 import { useFeature } from '../../features';
 
-const AdvancedNetworkFields: React.FC<{ isSNO: boolean }> = ({ isSNO }) => {
+type AdvancedNetworkFieldsProps = {
+  isSNO: boolean;
+  isIPv6: boolean;
+  isClusterCIDRIPv6: boolean;
+}
+
+const AdvancedNetworkFields: React.FC<AdvancedNetworkFieldsProps> = ({ isSNO, isIPv6, isClusterCIDRIPv6 }) => {
   const isNetworkTypeSelectionEnabled = useFeature(
     'ASSISTED_INSTALLER_NETWORK_TYPE_SELECTION_FEATURE',
   );
@@ -21,24 +26,11 @@ const AdvancedNetworkFields: React.FC<{ isSNO: boolean }> = ({ isSNO }) => {
     }
   };
 
-  const isClusterCIDRIPv6 = Address6.isValid(values.clusterNetworkCidr || '');
-  const isIPv6 = useMemo(
-    () =>
-      isClusterCIDRIPv6 ||
-      Address6.isValid(values.machineNetworkCidr || '') ||
-      Address6.isValid(values.serviceNetworkCidr || ''),
-    [isClusterCIDRIPv6, values.machineNetworkCidr, values.serviceNetworkCidr],
-  );
-
   const clusterNetworkHostPrefixHelperText = isClusterCIDRIPv6
     ? 'The subnet prefix length to assign to each individual node. For example, if Cluster Network Host Prefix is set to 116, then each node is assigned a /116 subnet out of the given cidr (clusterNetworkCIDR), which allows for 4,094 (2^(128 - 116) - 2) pod IPs addresses. If you are required to provide access to nodes from an external network, configure load balancers and routers to manage the traffic.'
     : 'The subnet prefix length to assign to each individual node. For example, if Cluster Network Host Prefix is set to 23, then each node is assigned a /23 subnet out of the given cidr (clusterNetworkCIDR), which allows for 510 (2^(32 - 23) - 2) pod IPs addresses. If you are required to provide access to nodes from an external network, configure load balancers and routers to manage the traffic.';
 
-  useEffect(() => {
-    if (isNetworkTypeSelectionEnabled && isIPv6) {
-      setFieldValue('networkType', NETWORK_TYPE_OVN);
-    }
-  }, [isIPv6, setFieldValue, isNetworkTypeSelectionEnabled]);
+
 
   return (
     <Grid hasGutter>
